start handling last version launched
