+++ src-tauri/Cargo.toml
@@ -36,3 +36,4 @@ reqwest = { version = "0.12.22"}
 base64 = "0.22.1"
 zip = "4.5.0"
 bytes = "1.10.1"
+semver = "1.0.26"
